+++ data/src/demo/java/ru/babaetskv/passionwoman/data/database/entity/transformations/ProductItemTransformableParamsProvider.kt
@@ -9,18 +9,18 @@ class ProductItemTransformableParamsProvider(
     private val database: PassionWomanDatabase
 ) : ProductItemEntity.TransformableParamsProvider {
 
-    override suspend fun provideImages(productItemId: Int): List<String> =
+    override suspend fun provideImages(productItemId: Long): List<String> =
         database.productImageDao.getForProductItem(productItemId)
             .map(AssetDbFormatter::formatAssetDbPath)
 
-    override suspend fun provideColor(colorId: Int): ColorModel =
+    override suspend fun provideColor(colorId: Long): ColorModel =
         database.colorDao.getById(colorId)
             ?.transform()
             ?: throw IllegalStateException("Failed to found color with id $colorId")
 
-    override suspend fun provideSizes(productItemId: Int): List<String>? =
+    override suspend fun provideSizes(productItemId: Long): List<String>? =
         database.productSizeDao.getForProductItem(productItemId).ifEmpty { null }
 
-    override suspend fun provideAvailableSizes(productItemId: Int): List<String>? =
+    override suspend fun provideAvailableSizes(productItemId: Long): List<String>? =
         database.productSizeDao.getAvailableForProductItem(productItemId).ifEmpty { null }
 }
