+++ data/src/demo/java/ru/babaetskv/passionwoman/data/preferences/FavoritesPreferencesImpl.kt
@@ -8,23 +8,23 @@ import java.util.*
 
 class FavoritesPreferencesImpl : FavoritesPreferences {
     private val actionsChannel = Channel<FavoritesPreferences.Action?>(Channel.RENDEZVOUS)
-    private val favoriteIds = LinkedList<Int>()
+    private val favoriteIds = LinkedList<Long>()
 
     override val favoritesUpdatesFlow: Flow<FavoritesPreferences.Action?>
         get() = actionsChannel.receiveAsFlow()
 
-    override fun isFavorite(id: Int): Boolean = favoriteIds.contains(id)
+    override fun isFavorite(id: Long): Boolean = favoriteIds.contains(id)
 
-    override fun putFavoriteId(id: Int) {
+    override fun putFavoriteId(id: Long) {
         if (favoriteIds.contains(id)) return
 
         favoriteIds.add(0, id)
         actionsChannel.trySend(FavoritesPreferences.Action.Put(id))
     }
 
-    override fun getFavoriteIds(): Collection<Int> = favoriteIds
+    override fun getFavoriteIds(): Collection<Long> = favoriteIds
 
-    override fun setFavoriteIds(ids: Collection<Int>) {
+    override fun setFavoriteIds(ids: Collection<Long>) {
         favoriteIds.run {
             clear()
             addAll(ids)
@@ -32,7 +32,7 @@ class FavoritesPreferencesImpl : FavoritesPreferences {
         actionsChannel.trySend(FavoritesPreferences.Action.Set)
     }
 
-    override fun setFavoriteIds(vararg ids: Int) {
+    override fun setFavoriteIds(vararg ids: Long) {
         favoriteIds.run {
             clear()
             addAll(ids.toList())
@@ -40,7 +40,7 @@ class FavoritesPreferencesImpl : FavoritesPreferences {
         actionsChannel.trySend(FavoritesPreferences.Action.Set)
     }
 
-    override fun deleteFavoriteId(id: Int) {
+    override fun deleteFavoriteId(id: Long) {
         if (!favoriteIds.contains(id)) return
 
         favoriteIds.run {
