+++ data/src/main/java/ru/babaetskv/passionwoman/data/gateway/ProfileGatewayImpl.kt
@@ -4,19 +4,16 @@ import android.net.Uri
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
 import okhttp3.RequestBody.Companion.asRequestBody
-import retrofit2.HttpException
 import ru.babaetskv.passionwoman.data.api.AuthApi
 import ru.babaetskv.passionwoman.data.gateway.base.BaseGatewayImpl
 import ru.babaetskv.passionwoman.data.model.ProfileModel
 import ru.babaetskv.passionwoman.domain.DateTimeConverter
 import ru.babaetskv.passionwoman.domain.StringProvider
-import ru.babaetskv.passionwoman.domain.exceptions.GatewayException
 import ru.babaetskv.passionwoman.domain.gateway.ProfileGateway
 import ru.babaetskv.passionwoman.domain.model.Order
 import ru.babaetskv.passionwoman.domain.model.Profile
 import ru.babaetskv.passionwoman.domain.model.base.Transformable
 import java.io.File
-import java.lang.Exception
 
 class ProfileGatewayImpl(
     private val api: AuthApi,
@@ -39,11 +36,11 @@ class ProfileGatewayImpl(
         }
     }
 
-    override suspend fun getFavoriteIds(): List<Int> = networkRequest {
+    override suspend fun getFavoriteIds(): List<Long> = networkRequest {
         api.getFavoriteIds()
     }
 
-    override suspend fun setFavoriteIds(ids: List<Int>) = networkRequest {
+    override suspend fun setFavoriteIds(ids: List<Long>) = networkRequest {
         api.setFavoriteIds(ids)
     }
