+++ domain/src/main/java/ru/babaetskv/passionwoman/domain/gateway/ProfileGateway.kt
@@ -10,7 +10,7 @@ interface ProfileGateway {
     suspend fun getProfile(): Transformable<Unit, Profile>
     suspend fun updateProfile(profile: Profile)
     suspend fun updateAvatar(imageUri: Uri)
-    suspend fun getFavoriteIds(): List<Int>
-    suspend fun setFavoriteIds(ids: List<Int>)
+    suspend fun getFavoriteIds(): List<Long>
+    suspend fun setFavoriteIds(ids: List<Long>)
     suspend fun getOrders(): List<Transformable<DateTimeConverter, Order>> // TODO: add pagination
 }
